Fix enqueue! typespec
It seems that dialyzer is not checking this typespec. Its probably
related to macros but I'm not entirely sure. `TaskBunny.Job.enqueue!`
spec clearly says that only return value is :ok.

This was discovered when using this library in an application:
```
lib/transporter/delivery_requests/delivery_job.ex:5: The specification
for 'Elixir.Transporter.DeliveryRequests.DeliveryJob':'enqueue!'/2
  states that the function might also return {'error',_} but the
  inferred return is 'ok'
```