Create CI github action workflow
Since this repository is forked from the original, the travis CI
workflow no longer works, therefore it has been removed and replaced
with Github Actions

This CI workflow uses the latest versions previously supported of
elixir 1.7 and erlang 21. It tests that the code is formatted, linted
and tests passes.

The CI workflow was necessary because running the tests
locally did not work reliable, whereas this solution does.